requirements.txt
streamlit==1.40.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
ultralytics==8.1.39
pytesseract==0.3.10
easyocr==1.7.1
python-dotenv==1.0.0